refactor: place files in one service folder
