+++ karman_app/lib/controllers/habit/habit_controller.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:karman_app/models/habits/habit.dart';
 import 'package:karman_app/models/habits/habit_log.dart';
-import 'package:karman_app/services/habit/habit_service.dart';
+import 'package:karman_app/services/habit_service.dart';
 import 'package:karman_app/services/notification_service.dart';
 import 'package:shared_preferences/shared_preferences.dart';
