+++ karman_app/lib/services/timer_service.dart
